chore: category item 불러오는 훅 분리
